refactor system functions
