+++ node/runtime/src/lib.rs
@@ -28,10 +28,10 @@ use primitives::utils::{
 };
 use storage::{StateDb, StateDbUpdate};
 use transaction::{
-    AsyncCall, Callback, CallbackInfo, CallbackResult, CreateAccountTransaction,
-    DeployContractTransaction, FunctionCallTransaction, LogEntry, ReceiptBody,
-    ReceiptTransaction, SendMoneyTransaction, SignedTransaction, StakeTransaction,
-    SwapKeyTransaction, TransactionBody, TransactionResult, TransactionStatus
+    AsyncCall, Callback, CallbackInfo, CallbackResult,
+    FunctionCallTransaction, LogEntry, ReceiptBody,
+    ReceiptTransaction, SignedTransaction,
+    TransactionBody, TransactionResult, TransactionStatus
 };
 use wasm::executor;
 use wasm::types::{ReturnData, RuntimeContext};
@@ -39,7 +39,10 @@ use chain::ReceiptBlock;
 
 use crate::ext::RuntimeExt;
 use crate::tx_stakes::{get_tx_stake_key, TxStakeConfig, TxTotalStake};
-use crate::system::{SYSTEM_METHOD_CREATE_ACCOUNT, SYSTEM_METHOD_DEPLOY, system_account};
+use crate::system::{
+    SYSTEM_METHOD_CREATE_ACCOUNT, SYSTEM_METHOD_DEPLOY, system_account,
+    system_create_account, system_deploy
+};
 
 pub mod test_utils;
 pub mod state_viewer;
@@ -166,171 +169,6 @@ impl Runtime {
         None
     }
 
-    fn send_money(
-        &self,
-        state_update: &mut StateDbUpdate,
-        transaction: &SendMoneyTransaction,
-        hash: CryptoHash,
-        sender: &mut Account,
-        accounting_info: AccountingInfo,
-    ) -> Result<Vec<ReceiptTransaction>, String> {
-        if transaction.amount == 0 {
-            return Err("Sending 0 tokens".to_string());
-        }
-        if sender.amount >= transaction.amount {
-            sender.amount -= transaction.amount;
-            set(state_update, &account_id_to_bytes(COL_ACCOUNT, &transaction.originator), sender);
-            let receipt = ReceiptTransaction::new(
-                transaction.originator.clone(),
-                transaction.receiver.clone(),
-                create_nonce_with_nonce(&hash, 0),
-                ReceiptBody::NewCall(AsyncCall::new(
-                    // Empty method name is used for deposit
-                    vec![],
-                    vec![],
-                    transaction.amount,
-                    0,
-                    accounting_info,
-                ))
-            );
-            Ok(vec![receipt])
-        } else {
-            Err(
-                format!(
-                    "Account {} tries to send {}, but has staked {} and only has {}",
-                    transaction.originator,
-                    transaction.amount,
-                    sender.staked,
-                    sender.amount,
-                )
-            )
-        }
-    }
-
-    fn staking(
-        &self,
-        state_update: &mut StateDbUpdate,
-        body: &StakeTransaction,
-        sender_account_id: &AccountId,
-        sender: &mut Account,
-        authority_proposals: &mut Vec<AuthorityStake>,
-    ) -> Result<Vec<ReceiptTransaction>, String> {
-        if sender.amount >= body.amount && sender.public_keys.is_empty() {
-            authority_proposals.push(AuthorityStake {
-                account_id: sender_account_id.clone(),
-                public_key: sender.public_keys[0],
-                amount: body.amount,
-            });
-            sender.amount -= body.amount;
-            sender.staked += body.amount;
-            set(state_update, &account_id_to_bytes(COL_ACCOUNT, sender_account_id), &sender);
-            Ok(vec![])
-        } else if sender.amount < body.amount {
-            let err_msg = format!(
-                "Account {} tries to stake {}, but has staked {} and only has {}",
-                body.originator,
-                body.amount,
-                sender.staked,
-                sender.amount,
-            );
-            Err(err_msg)
-        } else {
-            Err(format!("Account {} already staked", body.originator))
-        }
-    }
-
-    fn create_account(
-        &self,
-        state_update: &mut StateDbUpdate,
-        body: &CreateAccountTransaction,
-        hash: CryptoHash,
-        sender: &mut Account,
-        accounting_info: AccountingInfo,
-    ) -> Result<Vec<ReceiptTransaction>, String> {
-        if !is_valid_account_id(&body.new_account_id) {
-            return Err(format!("Account {} does not match requirements", body.new_account_id));
-        }
-        if sender.amount >= body.amount {
-            sender.amount -= body.amount;
-            set(
-                state_update,
-                &account_id_to_bytes(COL_ACCOUNT, &body.originator),
-                &sender
-            );
-            let new_nonce = create_nonce_with_nonce(&hash, 0);
-            let receipt = ReceiptTransaction::new(
-                body.originator.clone(),
-                body.new_account_id.clone(),
-                new_nonce,
-                ReceiptBody::NewCall(AsyncCall::new(
-                    SYSTEM_METHOD_CREATE_ACCOUNT.to_vec(),
-                    body.public_key.clone(),
-                    body.amount,
-                    0,
-                    accounting_info,
-                ))
-            );
-            Ok(vec![receipt])
-        } else {
-            Err(
-                format!(
-                    "Account {} tries to create new account with {}, but only has {}",
-                    body.originator,
-                    body.amount,
-                    sender.amount
-                )
-            )
-        }
-    }
-
-    fn swap_key(
-        &self,
-        state_update: &mut StateDbUpdate,
-        body: &SwapKeyTransaction,
-        account: &mut Account,
-    ) -> Result<Vec<ReceiptTransaction>, String> {
-        let cur_key = Decode::decode(&body.cur_key).map_err(|_| "cannot decode public key")?;
-        let new_key = Decode::decode(&body.new_key).map_err(|_| "cannot decode public key")?;
-        let num_keys = account.public_keys.len();
-        account.public_keys.retain(|&x| x != cur_key);
-        if account.public_keys.len() == num_keys {
-            return Err(format!("Account {} does not have public key {}", body.originator, cur_key));
-        }
-        account.public_keys.push(new_key);
-        set(
-            state_update,
-            &account_id_to_bytes(COL_ACCOUNT, &body.originator),
-            &account
-        );
-        Ok(vec![])
-    }
-
-    fn deploy(
-        &self,
-        body: &DeployContractTransaction,
-        hash: CryptoHash,
-        accounting_info: AccountingInfo,
-    ) -> Result<Vec<ReceiptTransaction>, String> {
-        // TODO: check signature
-        
-        let new_nonce = create_nonce_with_nonce(&hash, 0);
-        let args = Encode::encode(&(&body.public_key, &body.wasm_byte_array))
-            .map_err(|_| "cannot encode args")?;
-        let receipt = ReceiptTransaction::new(
-            body.originator.clone(),
-            body.contract_id.clone(),
-            new_nonce,
-            ReceiptBody::NewCall(AsyncCall::new(
-                SYSTEM_METHOD_DEPLOY.to_vec(),
-                args,
-                0,
-                0,
-                accounting_info,
-            ))
-        );
-        Ok(vec![receipt])
-    }
-
     fn call_function(
         &self,
         state_update: &mut StateDbUpdate,
@@ -425,7 +263,7 @@ impl Runtime {
                 ).ok_or_else(|| format!("sender {} does not have enough mana {}", sender_account_id, mana))?;
                 match transaction.body {
                     TransactionBody::SendMoney(ref t) => {
-                        self.send_money(
+                        system::send_money(
                             state_update,
                             &t,
                             transaction.get_hash(),
@@ -434,7 +272,7 @@ impl Runtime {
                         )
                     },
                     TransactionBody::Stake(ref t) => {
-                        self.staking(
+                        system::staking(
                             state_update,
                             &t,
                             &sender_account_id,
@@ -453,14 +291,14 @@ impl Runtime {
                         )
                     },
                     TransactionBody::DeployContract(ref t) => {
-                        self.deploy(
+                        system::deploy(
                             t,
                             transaction.get_hash(),
                             accounting_info,
                         )
                     },
                     TransactionBody::CreateAccount(ref t) => {
-                        self.create_account(
+                        system::create_account(
                             state_update,
                             t,
                             transaction.get_hash(),
@@ -469,7 +307,7 @@ impl Runtime {
                         )
                     },
                     TransactionBody::SwapKey(ref t) => {
-                        self.swap_key(
+                        system::swap_key(
                             state_update,
                             t,
                             &mut sender,
@@ -479,22 +317,6 @@ impl Runtime {
             }
             _ => Err(format!("sender {} does not exist", sender_account_id))
         }
-    }
-
-    fn deposit(
-        &self,
-        state_update: &mut StateDbUpdate,
-        amount: u64,
-        receiver_id: &AccountId,
-        receiver: &mut Account
-    ) -> Result<Vec<ReceiptTransaction>, String> {
-        receiver.amount += amount;
-        set(
-            state_update,
-            &account_id_to_bytes(COL_ACCOUNT, &receiver_id),
-            receiver
-        );
-        Ok(vec![])
     }
 
     fn return_data_to_receipts(
@@ -764,7 +586,7 @@ impl Runtime {
                             if amount > 0 {
                                 mana_accounting.mana_refund = async_call.mana;
                                 mana_accounting.accounting_info = async_call.accounting_info.clone();
-                                self.deposit(
+                                system::deposit(
                                     state_update,
                                     async_call.amount,
                                     &receipt.receiver,
@@ -869,14 +691,14 @@ impl Runtime {
                 if let ReceiptBody::NewCall(call) = &receipt.body {
                     amount = call.amount;
                     if call.method_name == SYSTEM_METHOD_CREATE_ACCOUNT {
-                        system::create_account(
+                        system_create_account(
                             state_update,
                             &call,
                             &receipt.receiver,
                         )
                     } else if call.method_name == SYSTEM_METHOD_DEPLOY {
                         // TODO(#413): Fix security of contract deploy.
-                        system::deploy(
+                        system_deploy(
                             state_update,
                             &call,
                             &receipt.receiver,
@@ -1145,24 +967,13 @@ mod tests {
     use std::sync::Arc;
 
     use primitives::hash::hash;
-    use primitives::signature::{get_key_pair, sign};
+    use primitives::signature::{get_key_pair};
     use storage::test_utils::create_state_db;
-    use transaction::TransactionBody;
 
     use crate::state_viewer::AccountViewCallResult;
     use crate::test_utils::*;
 
     use super::*;
-
-    fn alice_account() -> AccountId {
-        "alice.near".to_string()
-    }
-    fn bob_account() -> AccountId {
-        "bob.near".to_string()
-    }
-    fn eve_account() -> AccountId {
-        "eve.near".to_string()
-    }
 
     impl Default for Runtime {
         fn default() -> Runtime {
@@ -1178,11 +989,6 @@ mod tests {
                 state_db: self.state_db.clone()
             }
         }
-    }
-
-    fn default_code_hash() -> CryptoHash {
-        let genesis_wasm = include_bytes!("../../../core/wasm/runtest/res/wasm_with_mem.wasm");
-        hash(genesis_wasm)
     }
 
     // TODO(#348): Add tests for TX staking, mana charging and regeneration
@@ -1322,272 +1128,6 @@ mod tests {
         assert_eq!(apply_results[2].tx_result[0].status, TransactionStatus::Completed);
         // Checking final root
         assert_ne!(root, new_root);
-    }
-
-    #[test]
-    fn test_upload_contract() {
-        let (runtime, _viewer, root) = get_runtime_and_state_db_viewer();
-        let wasm_binary = include_bytes!("../../../core/wasm/runtest/res/wasm_with_mem.wasm");
-        let mut alice = User::new(runtime.clone(), &alice_account());
-        let (public_key, _) = get_key_pair();
-        let (new_root, mut apply_results) = alice.deploy_contract(
-            root, &eve_account(), public_key, wasm_binary
-        );
-        let apply_result = apply_results.pop().unwrap();
-        assert_eq!(apply_result.tx_result[0].status, TransactionStatus::Completed);
-        assert_eq!(apply_result.new_receipts.len(), 0);
-        assert_ne!(root, new_root);
-        let mut new_state_update = StateDbUpdate::new(runtime.state_db, new_root);
-        let code: Vec<u8> = get(
-            &mut new_state_update,
-            &account_id_to_bytes(COL_CODE, &eve_account())
-        ).unwrap();
-        assert_eq!(code, wasm_binary.to_vec());
-    }
-
-    #[test]
-    fn test_redeploy_contract() {
-        let test_binary = b"test_binary";
-        let (runtime, _, root) = get_runtime_and_state_db_viewer();
-        let mut state_update = StateDbUpdate::new(runtime.state_db.clone(), root);
-        let account: Account = get(
-            &mut state_update,
-            &account_id_to_bytes(COL_ACCOUNT, &bob_account())
-        ).unwrap();
-        let mut bob = User::new(runtime.clone(), &bob_account());
-        let (new_root, mut apply_results) = bob.deploy_contract(
-            root, &bob_account(), account.public_keys[0], test_binary
-        );
-        let apply_result = apply_results.pop().unwrap();
-        assert_eq!(apply_result.tx_result[0].status, TransactionStatus::Completed);
-        assert_eq!(apply_result.new_receipts.len(), 0);
-        assert_ne!(root, new_root);
-        let mut new_state_update = StateDbUpdate::new(runtime.state_db, new_root);
-        let code: Vec<u8> = get(
-            &mut new_state_update,
-            &account_id_to_bytes(COL_CODE, &bob_account())
-        ).unwrap();
-        assert_eq!(code, test_binary.to_vec())
-    }
-
-    #[test]
-    fn test_send_money() {
-        let (runtime, viewer, root) = get_runtime_and_state_db_viewer();
-        let mut alice = User::new(runtime, &alice_account());
-        let (new_root, apply_results) = alice.send_money(root, &bob_account(), 10);
-        for apply_result in apply_results {
-            assert_eq!(apply_result.tx_result[0].status, TransactionStatus::Completed);
-        }
-        assert_ne!(root, new_root);
-        let result1 = viewer.view_account(new_root, &alice_account());
-        assert_eq!(
-            result1.unwrap(),
-            AccountViewCallResult {
-                nonce: 1,
-                account: alice_account(),
-                amount: 90,
-                stake: 50,
-                code_hash: default_code_hash(),
-            }
-        );
-        let result2 = viewer.view_account(new_root, &bob_account());
-        assert_eq!(
-            result2.unwrap(),
-            AccountViewCallResult {
-                nonce: 0,
-                account: bob_account(),
-                amount: 10,
-                stake: 0,
-                code_hash: default_code_hash(),
-            }
-        );
-    }
-
-    #[test]
-    fn test_send_money_over_balance() {
-        let (runtime, viewer, root) = get_runtime_and_state_db_viewer();
-        let mut alice = User::new(runtime, &alice_account());
-        let (new_root, mut apply_results) = alice.send_money(root, &bob_account(), 1000);
-        let apply_result = apply_results.pop().unwrap();
-        assert_eq!(apply_result.tx_result[0].status, TransactionStatus::Failed);
-        assert_eq!(apply_result.new_receipts.len(), 0);
-        assert_eq!(root, new_root);
-        let result1 = viewer.view_account(new_root, &alice_account());
-        assert_eq!(
-            result1.unwrap(),
-            AccountViewCallResult {
-                nonce: 0,
-                account: alice_account(),
-                amount: 100,
-                stake: 50,
-                code_hash: default_code_hash(),
-            }
-        );
-        let result2 = viewer.view_account(apply_result.root, &bob_account());
-        assert_eq!(
-            result2.unwrap(),
-            AccountViewCallResult {
-                nonce: 0,
-                account: bob_account(),
-                amount: 0,
-                stake: 0,
-                code_hash: default_code_hash(),
-            }
-        );
-    }
-
-    #[test]
-    fn test_refund_on_send_money_to_non_existent_account() {
-        let (runtime, viewer, root) = get_runtime_and_state_db_viewer();
-        let mut alice = User::new(runtime, &alice_account());
-        let (new_root, apply_results) = alice.send_money(root, &eve_account(), 10);
-        // 3 results: signed tx, deposit receipt, refund
-        assert_eq!(apply_results.len(), 3);
-        assert_eq!(apply_results[0].tx_result[0].status, TransactionStatus::Completed);
-        // deposit receipt failed because account does not exist
-        assert_eq!(apply_results[1].tx_result[0].status, TransactionStatus::Failed);
-        assert_eq!(apply_results[2].tx_result[0].status, TransactionStatus::Completed);
-        // assert_eq!(apply_result.new_receipts.len(), 0);
-        let result1 = viewer.view_account(new_root, &alice_account());
-        assert_eq!(
-            result1.unwrap(),
-            AccountViewCallResult {
-                nonce: 1,
-                account: alice_account(),
-                amount: 100,
-                stake: 50,
-                code_hash: default_code_hash(),
-            }
-        );
-        let result2 = viewer.view_account(new_root, &eve_account());
-        assert!(result2.is_err());
-    }
-
-    #[test]
-    fn test_create_account() {
-        let (runtime, viewer, root) = get_runtime_and_state_db_viewer();
-        let mut alice = User::new(runtime, &alice_account());
-        let (new_root, _) = alice.create_account(root, &eve_account(), 10);
-        assert_ne!(root, new_root);
-        let result1 = viewer.view_account(new_root, &alice_account());
-        assert_eq!(
-            result1.unwrap(),
-            AccountViewCallResult {
-                nonce: 1,
-                account: alice_account(),
-                amount: 90,
-                stake: 50,
-                code_hash: default_code_hash(),
-            }
-        );
-        let result2 = viewer.view_account(new_root, &eve_account());
-        assert_eq!(
-            result2.unwrap(),
-            AccountViewCallResult {
-                nonce: 0,
-                account: eve_account(),
-                amount: 10,
-                stake: 0,
-                code_hash: hash(b""),
-            }
-        );
-    }
-
-    #[test]
-    fn test_create_account_failure_invalid_name() {
-        let (runtime, viewer, root) = get_runtime_and_state_db_viewer();
-        let mut alice = User::new(runtime, &alice_account());
-        for invalid_account_name in vec![
-                "eve", // too short
-                "Alice.near", // capital letter
-                "alice(near)", // brackets are invalid
-                "long_of_the_name_for_real_is_hard", // too long
-                "qq@qq*qq" // * is invalid
-        ] { 
-            let (new_root, _) = alice.create_account(root, invalid_account_name, 10);
-            assert_eq!(root, new_root);
-            let result1 = viewer.view_account(new_root, &alice_account());
-            assert_eq!(
-                result1.unwrap(),
-                AccountViewCallResult {
-                    nonce: 0,
-                    account: alice_account(),
-                    amount: 100,
-                    stake: 50,
-                    code_hash: default_code_hash(),
-                }
-            );
-        }
-    }
-
-    #[test]
-    fn test_create_account_failure_already_exists() {
-        let (runtime, viewer, root) = get_runtime_and_state_db_viewer();
-        let mut alice = User::new(runtime, &alice_account());
-        let (new_root, _) = alice.create_account(root, &bob_account(), 10);
-        assert_ne!(root, new_root);
-        let result1 = viewer.view_account(new_root, &alice_account());
-        assert_eq!(
-            result1.unwrap(),
-            AccountViewCallResult {
-                nonce: 1,
-                account: alice_account(),
-                amount: 100,
-                stake: 50,
-                code_hash: default_code_hash(),
-            }
-        );
-        let result2 = viewer.view_account(new_root, &bob_account());
-        assert_eq!(
-            result2.unwrap(),
-            AccountViewCallResult {
-                nonce: 0,
-                account: bob_account(),
-                amount: 0,
-                stake: 0,
-                code_hash: default_code_hash(),
-            }
-        );
-    }
-
-    #[test]
-    fn test_swap_key() {
-        let (mut runtime, _, root) = get_runtime_and_state_db_viewer();
-        let (pub_key1, secret_key1) = get_key_pair();
-        let (pub_key2, _) = get_key_pair();
-        let mut alice = User::new(runtime.clone(), &alice_account());
-        let (new_root, apply_results) = alice.create_account_with_key(
-            root, &eve_account(), 10, pub_key1
-        );
-        for apply_result in apply_results {
-            assert_eq!(apply_result.tx_result[0].status, TransactionStatus::Completed);
-        }
-        assert_ne!(root, new_root);
-        let tx_body = TransactionBody::SwapKey(SwapKeyTransaction {
-            nonce: 2,
-            originator: eve_account(),
-            cur_key: pub_key1.encode().unwrap(),
-            new_key: pub_key2.encode().unwrap(),
-        });
-        let data = tx_body.encode().unwrap();
-        let signature = sign(&data, &secret_key1);
-        let transaction1 = SignedTransaction::new(signature, tx_body);
-        let apply_state = ApplyState {
-            shard_id: 0,
-            root: new_root,
-            parent_block_hash: CryptoHash::default(),
-            block_index: 0,
-        };
-        let apply_result = runtime.apply(
-            &apply_state, &[], &[transaction1],
-        );
-        runtime.state_db.commit(apply_result.db_changes).unwrap();
-        let mut new_state_update = StateDbUpdate::new(runtime.state_db.clone(), apply_result.root);
-        let account = get::<Account>(
-            &mut new_state_update,
-            &account_id_to_bytes(COL_ACCOUNT, &eve_account()),
-        ).unwrap();
-        assert_eq!(account.public_keys, vec![pub_key2]);
     }
 
     #[test]
